persistent auth tokens can be issued and used for authentication

+++ src/client/java/netxms-client/src/main/java/org/netxms/client/AccessListElement.java
@@ -34,7 +34,7 @@ public class AccessListElement extends AbstractAccessListElement
 	 * @param userId user id
 	 * @param accessRights bit mask
 	 */
-	public AccessListElement(long userId, int accessRights)
+   public AccessListElement(int userId, int accessRights)
 	{
 		super(userId, accessRights);
       inherited = false;
