+++ src/client/nxmc/java/src/main/java/org/netxms/nxmc/modules/logviewer/widgets/UserConditionEditor.java
@@ -1,6 +1,6 @@
 /**
  * NetXMS - open source network management system
- * Copyright (C) 2003-2022 Victor Kirhenshtein
+ * Copyright (C) 2003-2024 Victor Kirhenshtein
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -72,7 +72,7 @@ public class UserConditionEditor extends ConditionEditor
       if ((initialFilter != null) && (initialFilter.getType() == ColumnFilterType.EQUALS))
       {
          setSelectedOperation(initialFilter.isNegated() ? 1 : 0);
-         userSelector.setUserId(initialFilter.getNumericValue());
+         userSelector.setUserId((int)initialFilter.getNumericValue());
       }
 	}
