+++ src/client/nxmc/java/src/main/java/org/netxms/nxmc/modules/objecttools/propertypages/AccessControl.java
@@ -1,6 +1,6 @@
 /**
  * NetXMS - open source network management system
- * Copyright (C) 2003-2020 Victor Kirhenshtein
+ * Copyright (C) 2003-2024 Victor Kirhenshtein
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -85,7 +85,7 @@ public class AccessControl extends PropertyPage
 	{	
 		// Build internal copy of access list
 		session = Registry.getSession();
-		for(Long uid : objectTool.getAccessList())
+      for(Integer uid : objectTool.getAccessList())
 		{
 			AbstractUserObject o = session.findUserDBObjectById(uid, null);
 			if (o != null)
@@ -185,13 +185,13 @@ public class AccessControl extends PropertyPage
          @Override
          protected void run(IProgressMonitor monitor) throws Exception
          {
-            if (session.syncMissingUsers(new HashSet<Long>(objectTool.getAccessList())))
+            if (session.syncMissingUsers(new HashSet<Integer>(objectTool.getAccessList())))
             {
                runInUIThread(new Runnable() {
                   @Override
                   public void run()
                   {    
-                     for(Long uid : objectTool.getAccessList())
+                     for(Integer uid : objectTool.getAccessList())
                      {
                         AbstractUserObject o = session.findUserDBObjectById(uid, null);
                         if (o != null)
@@ -246,7 +246,7 @@ public class AccessControl extends PropertyPage
 	{
 	   if (isControlCreated())
 	   {
-   		List<Long> list = new ArrayList<Long>(acl.size());
+         List<Integer> list = new ArrayList<Integer>(acl.size());
    		for(AbstractUserObject o : acl)
    			list.add(o.getId());
    		objectTool.setAccessList(list);
