+++ src/client/nxmc/java/src/main/java/org/netxms/nxmc/modules/users/propertypages/Members.java
@@ -1,6 +1,6 @@
 /**
  * NetXMS - open source network management system
- * Copyright (C) 2003-2021 Victor Kirhenshtein
+ * Copyright (C) 2003-2024 Victor Kirhenshtein
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -60,7 +60,7 @@ public class Members extends PropertyPage
    private TableViewer userList;
 	private NXCSession session;
 	private UserGroup object;
-	private HashMap<Long, AbstractUserObject> members = new HashMap<Long, AbstractUserObject>(0);
+   private HashMap<Integer, AbstractUserObject> members = new HashMap<Integer, AbstractUserObject>(0);
    
    /**
     * Default constructor
@@ -179,7 +179,7 @@ public class Members extends PropertyPage
       });
 
       // Initial data
-		for(long userId : object.getMembers())
+      for(int userId : object.getMembers())
 		{
 			final AbstractUserObject user = session.findUserDBObjectById(userId, null);
 			if (user != null)
@@ -202,10 +202,10 @@ public class Members extends PropertyPage
 	{
 		if (isApply)
 			setValid(false);
-		
-		long[] memberIds = new long[members.size()];
+
+      int[] memberIds = new int[members.size()];
 		int i = 0;
-		for(Long id : members.keySet())
+      for(Integer id : members.keySet())
 			memberIds[i++] = id;
 		object.setMembers(memberIds);
