+++ src/java/netxms-eclipse/AlarmViewer/src/org/netxms/ui/eclipse/alarmviewer/widgets/helpers/AlarmListLabelProvider.java
@@ -139,7 +139,7 @@ public class AlarmListLabelProvider extends LabelProvider implements ITableLabel
 			case AlarmList.COLUMN_ACK_BY:
             if (alarm.getState() == Alarm.STATE_OUTSTANDING)
 					return null;
-            long userId = (alarm.getState() == Alarm.STATE_ACKNOWLEDGED) ? alarm.getAcknowledgedByUser() : alarm.getResolvedByUser();
+            int userId = (alarm.getState() == Alarm.STATE_ACKNOWLEDGED) ? alarm.getAcknowledgedByUser() : alarm.getResolvedByUser();
             AbstractUserObject user = session.findUserDBObjectById(userId, new ViewerElementUpdater(viewer, element));
             return (user != null) ? user.getName() : ("[" + Long.toString(userId) + "]"); //$NON-NLS-1$ //$NON-NLS-2$
 			case AlarmList.COLUMN_CREATED:
