+++ webui/webapp/AlarmViewer/src/org/netxms/ui/eclipse/alarmviewer/propertypages/AccessControl.java
@@ -60,7 +60,7 @@ public class AccessControl extends PropertyPage
    private NXCSession session;
    private AlarmCategory category;
    private AlarmCategoryEditor editor;
-   private HashMap<Long, AbstractUserObject> accessMap = new HashMap<Long, AbstractUserObject>(0);
+   private HashMap<Integer, AbstractUserObject> accessMap = new HashMap<Integer, AbstractUserObject>(0);
 
    @Override
    protected Control createContents(Composite parent)
@@ -158,7 +158,7 @@ public class AccessControl extends PropertyPage
       });
 
       // Initial data
-      for(long userId : category.getAccessControl())
+      for(int userId : category.getAccessControl())
       {
          final AbstractUserObject user = session.findUserDBObjectById(userId, null);
          if (user != null)
@@ -187,13 +187,13 @@ public class AccessControl extends PropertyPage
          @Override
          protected void runInternal(IProgressMonitor monitor) throws Exception
          {
-            if (session.syncMissingUsers(new HashSet<Long>(Arrays.asList(category.getAccessControl()))))
+            if (session.syncMissingUsers(new HashSet<Integer>(Arrays.asList(category.getAccessControl()))))
             {
                runInUIThread(new Runnable() {
                   @Override
                   public void run()
                   {    
-                     for(long userId : category.getAccessControl())
+                     for(int userId : category.getAccessControl())
                      {
                         final AbstractUserObject user = session.findUserDBObjectById(userId, null);
                         if (user != null)
@@ -224,9 +224,9 @@ public class AccessControl extends PropertyPage
     */
    protected boolean applyChanges(final boolean isApply)
    {      
-      Long[] accessControlIds = new Long[accessMap.size()];
+      Integer[] accessControlIds = new Integer[accessMap.size()];
       int i = 0;
-      for(Long id : accessMap.keySet())
+      for(Integer id : accessMap.keySet())
          accessControlIds[i++] = id;
       category.setAccessControl(accessControlIds);
